+++ packages/core/src/internal/data-grid-dnd/data-grid-dnd.tsx
@@ -472,6 +472,7 @@ const DataGridDnd: React.FunctionComponent<DataGridDndProps> = p => {
             showFilter={p.showFilter}
             filterHeight={p.filterHeight}
             getFilterCellContent={p.getFilterCellContent}
+            getRowMarkerFilterCellContent={p.getRowMarkerFilterCellContent}
             dragCursor={dragCursor}
         />
     );