+++ packages/core/src/internal/data-grid/render/data-grid-render.ts
@@ -180,6 +180,8 @@ export function drawGrid(arg: DrawGridArg, lastArg: DrawGridArg | undefined) {
         filterHeight,
         showFilter,
         getFilterCellContent,
+        getRowMarkerFilterCellContent,
+        hasRowMarkers,
     } = arg;
     if (width === 0 || height === 0) return;
     const doubleBuffer = renderStrategy === "double-buffer";
@@ -314,7 +316,9 @@ export function drawGrid(arg: DrawGridArg, lastArg: DrawGridArg | undefined) {
             renderStateProvider,
             overrideCursor,
             getCellRenderer,
-            getFilterCellContent
+            getFilterCellContent,
+            getRowMarkerFilterCellContent,
+            hasRowMarkers
         );
 
         drawGridLines(