+++ packages/core/src/internal/scrolling-data-grid/scrolling-data-grid.tsx
@@ -353,6 +353,7 @@ const GridScroller: React.FunctionComponent<ScrollingDataGridProps> = p => {
                 showFilter={p.showFilter}
                 filterHeight={p.filterHeight}
                 getFilterCellContent={p.getFilterCellContent}
+                getRowMarkerFilterCellContent={p.getRowMarkerFilterCellContent}
             />
         </InfiniteScroller>
     );